requirements.txt
Flask==3.0.0
Pillow==10.1.0
pytesseract==0.3.10
rapid-latex-ocr==0.0.6
